Changed type from int to float for CONNECT and READ timeouts in api.py in order to be able to be more flexible when choosing seconds

+++ config/api.py
@@ -24,5 +24,5 @@ MAX_PAGES = int(os.getenv("MAX_PAGES"))
 MAX_ATTEMPTS = int(os.getenv("MAX_ATTEMPTS"))
 
 # Connection and reading request times
-CONNECT_TIMEOUT = int(os.getenv("CONNECT_TIMEOUT"))
-READ_TIMEOUT = int(os.getenv("READ_TIMEOUT"))
+CONNECT_TIMEOUT = float(os.getenv("CONNECT_TIMEOUT"))
+READ_TIMEOUT = float(os.getenv("READ_TIMEOUT"))
